1. The last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -14,7 +14,7 @@ int main(void)
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
 
-	long int l;
+	int l;
 	
 	l = n % 10;
